+++ app/javascript/types/index.js
@@ -1,0 +1,2 @@
+export const REQUEST_RECIPES='REQUEST_RECIPES'
+export const RECEIVE_RECIPES='RECEIVE_RECIPES'
